More clippy

+++ ceno_zkvm/src/instructions/riscv/logic/test.rs
@@ -155,7 +155,7 @@ fn test_opcode_xor() {
     let lkm = lkm.into_finalize_result()[ROMType::Xor as usize].clone();
     assert_eq!(&lkm.into_iter().sorted().collect_vec(), LOOKUPS);
 
-    let expected_rd_written = UInt8::from_const_unchecked(split_to_u8::<u64>((A ^ B) as u32));
+    let expected_rd_written = UInt8::from_const_unchecked(split_to_u8::<u64>(A ^ B));
 
     config
         .rd_written
